refactor(types): category 및 product 타입 분리 및 파일 정리
- Category, Product 인터페이스를 각각 types/category.ts, types/product.ts로 분리하였습니다.
- mockCategories, mockProducts는 기존 위치 유지하며 타입 import 방식으로 변경하였습니다.
