+++ src/types/category.ts
@@ -1,0 +1,5 @@
+export interface Category {
+  themeId: number;
+  name: string;
+  image: string;
+}
